refactor: Simplify data layer package structure
- Move all files from data/saved_recipe/ to data/
- Update package declarations and imports
- Remove redundant saved_recipe directory
