+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/data_source/RecipeDatasource.kt
@@ -1,0 +1,7 @@
+package com.surivalcoding.composerecipeapp.data.data_source
+
+import com.surivalcoding.composerecipeapp.data.dto.RecipeDto
+
+interface RecipeDatasource {
+    suspend fun getSavedRecipes(): List<RecipeDto>
+}
