+++ app/src/main/java/com/surivalcoding/composerecipeapp/data/dto/IngredientAmountDto.kt
@@ -1,4 +1,4 @@
-package com.surivalcoding.composerecipeapp.data.saved_recipe.dto
+package com.surivalcoding.composerecipeapp.data.dto
 
 import kotlinx.serialization.Serializable
