--- app/src/main/java/com/surivalcoding/composerecipeapp/data/saved_recipe/repository/RecipeRepository.kt
@@ -1,7 +1,0 @@
-package com.surivalcoding.composerecipeapp.data.saved_recipe.repository
-
-import com.surivalcoding.composerecipeapp.data.saved_recipe.model.Recipe
-
-interface RecipeRepository {
-    suspend fun getSavedRecipes(): List<Recipe>
-}
